ROX-34779: skip DeleteOldResults on scan watcher timeout (#20879)

+++ central/complianceoperator/v2/report/manager/manager_impl.go
@@ -491,8 +491,12 @@ func (m *managerImpl) handleReadyScan() {
 				Observe(timeActive.Minutes())
 			delete(m.watchingScansStartTime, scanWatcherResult.WatcherID)
 		})
-		if err := watcher.DeleteOldResults(m.automaticReportingCtx, scanWatcherResult, m.checkResultDataStore); err != nil {
-			log.Errorf("unable to delete old CheckResults: %v", err)
+		if scanWatcherResult.Error == nil || errors.Is(scanWatcherResult.Error, watcher.ErrScanRemoved) {
+			if err := watcher.DeleteOldResults(m.automaticReportingCtx, scanWatcherResult, m.checkResultDataStore); err != nil {
+				log.Errorf("unable to delete old CheckResults: %v", err)
+			}
+		} else {
+			log.Debugf("Skipping DeleteOldResults for scan %s: %v", scanWatcherResult.Scan.GetScanName(), scanWatcherResult.Error)
 		}
 		if errors.Is(scanWatcherResult.Error, watcher.ErrScanRemoved) {
 			log.Debugf("Scan %s was removed", scanWatcherResult.Scan.GetScanName())
